Rename collider renderer files to geometry renderer
