+++ Sprocket/CMakeLists.txt
@@ -62,7 +62,7 @@ add_library(Sprocket STATIC
             Graphics/texture.cpp
             Graphics/Viewport.cpp
 
-            Graphics/Rendering/ColliderRenderer.cpp
+            Graphics/Rendering/geometry_renderer.cpp
             Graphics/Rendering/Scene3DRenderer.cpp
             Graphics/Rendering/SkyboxRenderer.cpp
